backend/requirements.txt
# Web框架
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# 数据库
sqlalchemy==2.0.25
pymysql==1.1.0
alembic==1.13.1

# Redis
redis==5.0.1

# 数据处理
pandas==2.2.0
numpy==1.26.3

# 数据源
akshare==1.13.0
tushare==1.4.0

# QMT接口
xtquant==1.0.0

# 工具
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
httpx==0.26.0
pyjwt>=2.8.0
bcrypt>=4.1.0

# 异步
aiofiles==23.2.1
